no warning

+++ kernel/libc/include/stdio.h
@@ -16,8 +16,8 @@ extern struct FILE *stderr;
 extern struct FILE *stdout;
 extern struct FILE *stdin;
 
-size_t snprintf (char *, size_t, const char *, ...);
-size_t vsnprintf (char *, size_t, const char *, va_list);
+int snprintf (char *, size_t, const char *, ...);
+int vsnprintf (char *, size_t, const char *, va_list);
 int asprintf (char **, const char *, ...);
 
 int fprintf(struct FILE *f, const char * fmt, ...);
